aggregator/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.0
asyncpg==0.29.0
pydantic-settings==2.4.0
paho-mqtt==1.6.1
httpx==0.27.2
prometheus-client==0.21.0
